+++ .github/CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.1.0] - 2020-12-28
+### Added
+- `type()` method to `FakerResolver::class`. This allows you to specify the type that the generated value should equal.
+
 ## [v1.0.1] - 2020-12-28
 ### Added
-- Type hinting for FakerResolver class.
+- Type hinting for `FakerResolver::class`.
 
 ### Fixed
 - "Address" in dictionary was returning "safeAddress" which does not exist on Faker.
@@ -26,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
+[v1.0.1]: https://github.com/Anteris-Dev/faker-map/compare/v1.0.1...v1.1.0
 [v1.0.1]: https://github.com/Anteris-Dev/faker-map/compare/v1.0.0...v1.0.1
 [v1.0.0]: https://github.com/Anteris-Dev/faker-map/compare/v0.1.0...v1.0.0
 [v0.1.0]: https://github.com/Anteris-Dev/faker-map/releases/tag/v0.1.0